fix(location): prevent deletion of a location if it's associated with a FA or a FT #2321

+++ apps/api/src/location/location.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { ForbiddenException, Injectable } from "@nestjs/common";
 import { PrismaService } from "../prisma.service";
 import { CreateLocationRequestDto } from "./dto/create-location.request.dto";
 import { UpdateLocationRequestDto } from "./dto/update-location.request.dto";
@@ -26,7 +26,31 @@ export class LocationService {
     });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const [linkedActivities, linkedTasks] = await Promise.all([
+      this.prisma.festivalActivity.findMany({
+        where: { locationId: id },
+        select: { id: true },
+      }),
+      this.prisma.festivalTask.findMany({
+        where: { appointmentId: id },
+        select: { id: true },
+      }),
+    ]);
+
+    if (linkedActivities.length > 0 || linkedTasks.length > 0) {
+      const activitiesError = linkedActivities
+        .map((activity) => `FA ${activity.id}`)
+        .join(", ");
+      const tasksError = linkedTasks.map((task) => `FT ${task.id}`).join(", ");
+      const separator =
+        linkedActivities.length > 0 && linkedTasks.length > 0 ? ", " : "";
+
+      throw new ForbiddenException(
+        `Impossible de supprimer le lieu, il est lié à : ${activitiesError}${separator}${tasksError}`,
+      );
+    }
+
     return this.prisma.signaLocation.delete({ where: { id } });
   }
 }
